Ajuste limpeza de backups

+++ recipes/clean_backups.rb
@@ -24,6 +24,7 @@ bash 'clean_backups' do
     find #{DIR_BACKUP} -type f -name "#{PREFIX_D}*.#{EXT}" -mtime +#{RET_D} -exec rm -f {} \\;
     find #{DIR_BACKUP} -type f -name "#{PREFIX_M}*.#{EXT}" -mtime +#{RET_M} -exec rm -f {} \\;
     CODE
+    only_if "[[ -z $(ps -ef | grep 'chef-server-ctl backup --yes' | grep -v grep) ]]"
   end
 
 #bash 'clean_diario' do
